requirements.txt
streamlit
tqdm~=4.64.1
numpy~=1.21.6
torch~=1.12.1
adapter-transformers==2.2.0
datasets~=2.7.0
rouge_score~=0.1.2
jsonlines~=3.1.0
pandas~=1.3.5
python-dotenv